backend/payments/services.py
import uuid
import requests
from django.conf import settings
from .models import Payment

class BasePaymentService:
    def __init__(self, payment):
        self.payment = payment
        self.transaction_id = str(uuid.uuid4())

    def create_payment(self):
        raise NotImplementedError

    def verify_payment(self):
        raise NotImplementedError

class ClickPaymentService(BasePaymentService):
    def create_payment(self):
        # Implement Click payment creation
        # This is a placeholder for actual Click API integration
        return {
            'payment_url': f'https://click.uz/pay/{self.transaction_id}',
            'transaction_id': self.transaction_id
        }

    def verify_payment(self):
        # Implement Click payment verification
        # This is a placeholder for actual Click API verification
        return True

class PaymePaymentService(BasePaymentService):
    def create_payment(self):
        # Implement Payme payment creation
        # This is a placeholder for actual Payme API integration
        return {
            'payment_url': f'https://payme.uz/pay/{self.transaction_id}',
            'transaction_id': self.transaction_id
        }

    def verify_payment(self):
        # Implement Payme payment verification
        # This is a placeholder for actual Payme API verification
        return True

class UzumPaymentService(BasePaymentService):
    def create_payment(self):
        # Implement Uzum Bank payment creation
        # This is a placeholder for actual Uzum Bank API integration
        return {
            'payment_url': f'https://uzum.uz/pay/{self.transaction_id}',
            'transaction_id': self.transaction_id
        }

    def verify_payment(self):
        # Implement Uzum Bank payment verification
        # This is a placeholder for actual Uzum Bank API verification
        return True

def get_payment_service(payment):
    services = {
        'click': ClickPaymentService,
        'payme': PaymePaymentService,
        'uzum': UzumPaymentService,
    }
    service_class = services.get(payment.payment_method)
    if not service_class:
        raise ValueError(f"Unsupported payment method: {payment.payment_method}")
    return service_class(payment) 