backend/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
openai==1.69.0
python-dotenv==1.1.0
Pillow==10.2.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34 